+++ picamclient.py
@@ -14,14 +14,17 @@ from time import sleep
 from utils import ThreadEvents
 from config import Config,CheckConfig
 from threading import Semaphore
-from picam2 import ImageCapture, VideoCapture, HTTPStreamCapture
+from picam2 import ImageCapture, VideoCapture,VideoCaptureElapse, HTTPStreamCapture
 from picam2 import getCameraInfo, UDPStreamCapture
 from pantilt.ULN2003.stepmotors import PanTiltStepMotors
 from pantilt.waveshare.servomotors import PanTiltServoMotors
 from pantilt.waveshare.lightsensor import LightSensor
 
 MQTT_CLIENT_ID = 'picam2ctrl'
-
+TIMER_MIN = 1
+RATE_MIN = 1
+RATE_MAX = 50
+CNT_MIN = 0
 """
 QOS: 0 => fire and forget A -> B
 QOS: 1 => at leat one - msg will be send
@@ -43,9 +46,15 @@ LOG_LEVEL = {
     "WARN": logging.WARN,
     "DEBUG": logging.DEBUG
 }
+TP_REC = "Recording"
 TP_SNAP = "Snapshot"
+TP_SNAPCNT = "SnapCounter"
+TP_SNAPTI = "SnapTimer"
 TP_VIDEO = "Video"
+TP_VIDEOTI = "VideoTimer"
+TP_VIDSPEED = "VidLapseSpeed"
 TP_HTTP = "HttpStream"
+TP_TIMELAPSE = "TimeLapse"
 TP_UDP = "UdpStream"
 TP_MOTION = "Motion"
 TP_MOTION_EN = "MotionEnabled"
@@ -93,18 +102,24 @@ class PiCam2Client (hass.MQTTClient):
         get a dict() which defines the required client topic
         based on HASS types
         """
-        clientTopics = {TP_SNAP: hass.HASS_COMPONENT_SWITCH,
-                              TP_VIDEO: hass.HASS_COMPONENT_SWITCH,
-                              TP_HTTP: hass.HASS_COMPONENT_SWITCH,
-                              TP_UDP: hass.HASS_COMPONENT_SWITCH,
-                              TP_MOTION_EN: hass.HASS_COMPONENT_SWITCH,
-                              TP_MOTION: hass.HASS_COMPONENT_BINARY_SENSOR,
-                              TP_PANA: hass.HASS_COMPONENT_SWITCH,
-                              TP_PAN: hass.HASS_COMPONENT_NUMBER,
-                              TP_TILT: hass.HASS_COMPONENT_NUMBER,
-                              TP_LSENS: hass.HASS_COMPONENT_SENSOR}
-        
-        
+
+        clientTopics = {TP_REC: hass.HASS_COMPONENT_BINARY_SENSOR,
+                        TP_SNAP: hass.HASS_COMPONENT_SWITCH,
+                        TP_SNAPTI: hass.HASS_COMPONENT_NUMBER,
+                        TP_SNAPCNT: hass.HASS_COMPONENT_NUMBER,
+                        TP_VIDEO: hass.HASS_COMPONENT_SWITCH,
+                        TP_VIDEOTI: hass.HASS_COMPONENT_NUMBER,
+                        TP_VIDSPEED: hass.HASS_COMPONENT_NUMBER,
+                        TP_TIMELAPSE: hass.HASS_COMPONENT_SWITCH,
+                        TP_HTTP: hass.HASS_COMPONENT_SWITCH,
+                        TP_UDP: hass.HASS_COMPONENT_SWITCH,
+                        TP_MOTION_EN: hass.HASS_COMPONENT_SWITCH,
+                        TP_MOTION: hass.HASS_COMPONENT_BINARY_SENSOR,
+                        TP_PANA: hass.HASS_COMPONENT_SWITCH,
+                        TP_PAN: hass.HASS_COMPONENT_NUMBER,
+                        TP_TILT: hass.HASS_COMPONENT_NUMBER,
+                        TP_LSENS: hass.HASS_COMPONENT_SENSOR}
+
         return self.removeTpsByHW(clientTopics)
 
     def removeTpsByHW(self,tps:dict) -> dict:
@@ -126,43 +141,68 @@ class PiCam2Client (hass.MQTTClient):
         get a dict() which defines the required config topics
         based on HASS
         """
-        hassconfigs = dict()
+        hassconfigs = {
+            TP_REC: {hass.HASS_CONFIG_DEVICE_CLASS : hass.HASS_CLASS_RUNNING,
+                     hass.HASS_CONFIG_PAYLOAD_ON : True,
+                     hass.HASS_CONFIG_PAYLOAD_OFF : False},
+            TP_SNAP: {hass.HASS_CONFIG_DEVICE_CLASS : hass.HASS_CLASS_SWITCH,
+                      hass.HASS_CONFIG_ICON:"mdi:camera"},
+            TP_TIMELAPSE:{hass.HASS_CONFIG_DEVICE_CLASS : hass.HASS_CLASS_SWITCH,
+                          hass.HASS_CONFIG_ICON:"mdi:timelapse"},
+            TP_VIDEO: {hass.HASS_CONFIG_DEVICE_CLASS : hass.HASS_CLASS_SWITCH,
+                       hass.HASS_CONFIG_ICON:"mdi:file-video"},
+            TP_VIDSPEED: {
+                        #hass.HASS_CONFIG_ICON:"mdi:speedometer",
+                        hass.HASS_CONFIG_DEVICE_CLASS : hass.HASS_CLASS_SPEED,
+                        #hass.HASS_CONFIG_VALUE_TEMPLATE :"{{ value }}",
+                        #hass.HASS_CONFIG_CMD_TEMPLATE :"{{ value }}",
+                        hass.HASS_CONFIG_MIN : 2,
+                        hass.HASS_CONFIG_MAX : 20,
+                        hass.HASS_CONFIG_STEP : 1,
+                        hass.HASS_CONFIG_MODE : "slider",
+                        hass.HASS_CONFIG_CMD_TP: self._subTopics[TP_VIDSPEED]},
+            TP_HTTP: {hass.HASS_CONFIG_DEVICE_CLASS : hass.HASS_CLASS_SWITCH,
+                      hass.HASS_CONFIG_ICON:"mdi:video"},
+            TP_UDP: {hass.HASS_CONFIG_DEVICE_CLASS : hass.HASS_CLASS_SWITCH,
+                     hass.HASS_CONFIG_ICON:"mdi:video"},
+            TP_MOTION_EN: {hass.HASS_CONFIG_DEVICE_CLASS : hass.HASS_CLASS_SWITCH,
+                           hass.HASS_CONFIG_ICON:"mdi:video"},
+            TP_MOTION: {hass.HASS_CONFIG_ICON:"mdi:motion-sensor",
+                        hass.HASS_CONFIG_DEVICE_CLASS : hass.HASS_CLASS_MOTION},
+                        #hass.HASS_CONFIG_ATTR : f"{self.baseTopic}/{TP_MOTION}"},
+            TP_SNAPTI: {hass.HASS_CONFIG_ICON:"mdi:av-timer",
+                        hass.HASS_CONFIG_DEVICE_CLASS : hass.HASS_CLASS_DURATION,
+                        #hass.HASS_CONFIG_VALUE_TEMPLATE :"{{ value }}",
+                        #hass.HASS_CONFIG_CMD_TEMPLATE :"{{ value }}",
+                        hass.HASS_CONFIG_UNIT : "s",
+                        hass.HASS_CONFIG_MIN : TIMER_MIN,
+                        hass.HASS_CONFIG_MAX : 120,
+                        hass.HASS_CONFIG_STEP : 1,
+                        hass.HASS_CONFIG_MODE : "box",
+                        hass.HASS_CONFIG_CMD_TP: self._subTopics[TP_SNAPTI]},
+            TP_VIDEOTI:{hass.HASS_CONFIG_ICON:"mdi:av-timer",
+                        hass.HASS_CONFIG_DEVICE_CLASS : hass.HASS_CLASS_DURATION,
+                        #hass.HASS_CONFIG_VALUE_TEMPLATE :"{{ value }}",
+                        #hass.HASS_CONFIG_CMD_TEMPLATE :"{{ value }}",
+                        hass.HASS_CONFIG_UNIT : "s",
+                        hass.HASS_CONFIG_MIN : TIMER_MIN,
+                        hass.HASS_CONFIG_MAX : 7200,
+                        hass.HASS_CONFIG_STEP : 1,
+                        hass.HASS_CONFIG_MODE : "box",
+                        hass.HASS_CONFIG_CMD_TP: self._subTopics[TP_VIDEOTI]},
+            TP_SNAPCNT:{hass.HASS_CONFIG_ICON:"mdi:counter",
+                       #HASS device class counter unknown
+                       hass.HASS_CONFIG_DEVICE_CLASS : hass.HASS_CLASS_NONE,
+                       #hass.HASS_CONFIG_VALUE_TEMPLATE :"{{ value }}",
+                       #hass.HASS_CONFIG_CMD_TEMPLATE :"{{ value }}",
+                       hass.HASS_CONFIG_MIN : CNT_MIN,
+                       hass.HASS_CONFIG_MAX : 200,
+                       hass.HASS_CONFIG_STEP : 1,
+                       hass.HASS_CONFIG_MODE : "box",
+                       hass.HASS_CONFIG_CMD_TP: self._subTopics[TP_SNAPCNT]}
+            }
         for tp in self.setupClientTopics(): # depends on configured HW
-            if TP_SNAP == tp:
-                hassconfigs.update({TP_SNAP: 
-                                    {hass.HASS_CONFIG_ICON:"mdi:camera",
-                                     hass.HASS_CONFIG_DEVICE_CLASS : hass.HASS_CLASS_SWITCH}
-                                   })
-            
-            elif TP_VIDEO == tp:
-                hassconfigs.update({TP_VIDEO: 
-                                    {hass.HASS_CONFIG_ICON:"mdi:file-video",
-                                     hass.HASS_CONFIG_DEVICE_CLASS : hass.HASS_CLASS_SWITCH}
-                                   })
-            elif TP_HTTP == tp:
-                hassconfigs.update({TP_HTTP:
-                                    {hass.HASS_CONFIG_ICON:"mdi:video",
-                                     hass.HASS_CONFIG_DEVICE_CLASS : hass.HASS_CLASS_SWITCH}
-                                    })
-                
-            elif TP_UDP == tp:
-                hassconfigs.update({TP_UDP:
-                                    {hass.HASS_CONFIG_ICON:"mdi:video",
-                                     hass.HASS_CONFIG_DEVICE_CLASS : hass.HASS_CLASS_SWITCH}
-                                    })
-                
-            elif TP_MOTION_EN == tp:
-                hassconfigs.update({TP_MOTION_EN:
-                                    {hass.HASS_CONFIG_ICON:"mdi:video",
-                                     hass.HASS_CONFIG_DEVICE_CLASS : hass.HASS_CLASS_SWITCH}
-                                    })
-            elif TP_MOTION == tp:
-                hassconfigs.update({TP_MOTION:
-                                    {hass.HASS_CONFIG_ICON:"mdi:motion-sensor",
-                                     hass.HASS_CONFIG_DEVICE_CLASS : hass.HASS_CLASS_MOTION,
-                                     hass.HASS_CONFIG_ATTR : f"{self.baseTopic}/{TP_MOTION}"}
-                                    })
-            elif TP_PAN == tp:
+            if TP_PAN == tp:
                 hassconfigs.update({TP_PAN:
                                     {hass.HASS_CONFIG_ICON:"mdi:pan-horizontal",
                                      hass.HASS_CONFIG_DEVICE_CLASS : hass.HASS_CLASS_DISTANCE,
@@ -190,8 +230,8 @@ class PiCam2Client (hass.MQTTClient):
                                     })
             elif TP_PANA == tp:
                 hassconfigs.update({TP_PANA:
-                                    {hass.HASS_CONFIG_ICON:"mdi:pan-horizontal",
-                                     hass.HASS_CONFIG_DEVICE_CLASS : hass.HASS_CLASS_SWITCH}
+                                    {hass.HASS_CONFIG_DEVICE_CLASS : hass.HASS_CLASS_SWITCH,
+                                     hass.HASS_CONFIG_ICON:"mdi:pan-horizontal"}
                                      })
             elif TP_LSENS == tp:
                 hassconfigs.update({TP_LSENS:
@@ -200,8 +240,6 @@ class PiCam2Client (hass.MQTTClient):
                                      hass.HASS_CONFIG_VALUE_TEMPLATE :"{{ value }}",
                                      hass.HASS_CONFIG_CMD_TEMPLATE :"{{ value }}"}
                                      })
-            else:
-                logging.error(f"undefined tp_hass_config:{tp}")
         
         return hassconfigs
 
@@ -241,25 +279,39 @@ class PiCam2Client (hass.MQTTClient):
 
         for tp in self.TopicValues:
             val = self.TopicValues[tp]
-            if TP_SNAP == tp:
-                val= "ON" if self._snapshot else "OFF"
+            if TP_REC == tp:
+                val= False
+            elif TP_SNAP == tp:
+                val= hass.HASS_STATE_ON if self._snapshot else hass.HASS_STATE_OFF
             elif TP_VIDEO == tp:
-                val= "ON" if self._video else "OFF"
+                val= hass.HASS_STATE_ON if self._video else hass.HASS_STATE_OFF
+            elif TP_TIMELAPSE == tp:
+                val= hass.HASS_STATE_OFF
+            elif TP_VIDSPEED == tp:
+                val= 10
             elif TP_HTTP == tp:
-                val= "ON" if self._hstream else "OFF"
+                val= hass.HASS_STATE_ON if self._hstream else hass.HASS_STATE_OFF
             elif TP_UDP == tp:
-                val= "ON" if self._ustream else "OFF"
+                val= hass.HASS_STATE_ON if self._ustream else hass.HASS_STATE_OFF
             elif TP_MOTION_EN == tp:
-                val= "ON" if self._motionEnabled else "OFF"
+                val= hass.HASS_STATE_ON if self._motionEnabled else hass.HASS_STATE_OFF
             elif TP_MOTION == tp:
                 val = False
             elif TP_PANA == tp and self._PanTiltCam:
-                val= "ON" if self._PanTiltCam.get_Pana_active() else "OFF"
+                val= hass.HASS_STATE_ON if self._PanTiltCam.get_Pana_active() else hass.HASS_STATE_OFF
             elif TP_PAN == tp:
                 val = self._panAngle
             elif TP_TILT == tp:
                 val = self._tiltAngle
-            
+            elif TP_SNAPTI == tp:
+                val = 10 #every 10s
+            elif TP_VIDEOTI == tp:
+                val = 60# duration in s
+            elif TP_TILT == tp:
+                val = self._tiltAngle
+            elif TP_SNAPCNT == tp:
+                val = 1 # init value , 0 = endöess
+
             self.TopicValues[tp]= val
 
         mqtt_device = {
@@ -323,21 +375,21 @@ class PiCam2Client (hass.MQTTClient):
                     logging.debug(f"Camera PAN Auto Motion {payload}")
                     if self._PanTiltCam:
                         self.TopicValues[tp]= payload
-                        if payload == "ON" and not self._PanTiltCam.get_Pana_active():
+                        if payload == hass.HASS_STATE_ON and not self._PanTiltCam.get_Pana_active():
                             active=True
-                        elif payload == "OFF" and self._PanTiltCam.get_Pana_active():
+                        elif payload == hass.HASS_STATE_OFF and self._PanTiltCam.get_Pana_active():
                             active=False
                         self._PanTiltCam.pan_rotate_auto(active)
                         self.publish_state(self._stTopics[tp], payload= payload)
                     break
                 elif TP_MOTION_EN == tp:
                     self.TopicValues[tp]= payload
-                    if payload == "ON" and not self._child:
+                    if payload == hass.HASS_STATE_ON and not self._child:
                         #ignore since child app is active
                         self._motionEnabled = True
                         self.publish_state(TP_MOTION_EN)
                         self.publish_avail(self._avTopics[TP_MOTION])
-                    elif payload == "OFF":
+                    elif payload == hass.HASS_STATE_OFF:
                         self._motionEnabled = False
                         self.publish_state(TP_MOTION_EN)
                         self.publish_avail(self._avTopics[TP_MOTION], False)
@@ -347,35 +399,75 @@ class PiCam2Client (hass.MQTTClient):
                 elif not self._child:
                     if TP_SNAP == tp:
                         self.TopicValues[tp]= payload
-                        if payload == "ON":
-                            self._child = ImageCapture(self, self.cfg, self._motionEnabled)
+                        if payload == hass.HASS_STATE_ON:
+                            self._child = ImageCapture(self, self.cfg,
+                                                       self.TopicValues[TP_SNAPTI],
+                                                       self.TopicValues[TP_SNAPCNT],
+                                                       self.TopicValues[TP_TIMELAPSE]==hass.HASS_STATE_ON,
+                                                       self._motionEnabled)
                             self._snapshot = True
-                            self.publish_state(TP_SNAP)
                             self.activeThreads.addThread(self._child)
+                            self.publish_state(TP_SNAP)
+                            self.TopicValues[TP_REC]=True
+                            self.publish_state(TP_REC)
+                        break
+                    elif TP_SNAPTI == tp:
+                        self.TopicValues[tp]= payload
+                        self.publish_state(TP_SNAPTI)
+                        break
+                    elif TP_SNAPCNT == tp:
+                        self.TopicValues[tp]= payload
+                        self.publish_state(TP_SNAPCNT)
+                        break
+                    elif TP_VIDEOTI == tp:
+                        self.TopicValues[tp]= payload
+                        self.publish_state(TP_VIDEOTI)
+                        break
+                    elif TP_TIMELAPSE == tp:
+                        self.TopicValues[tp]= payload
+                        self.publish_state(TP_TIMELAPSE)
+                        break
+                    elif TP_VIDSPEED == tp:
+                        self.TopicValues[tp]= payload
+                        self.publish_state(TP_VIDSPEED)
                         break
                     elif TP_VIDEO == tp:
                         self.TopicValues[tp]= payload
-                        if payload == "ON":
-                            self._child = VideoCapture(self, self.cfg, self._motionEnabled)
+                        if payload == hass.HASS_STATE_ON:
+                            if hass.HASS_STATE_ON==self.TopicValues[TP_TIMELAPSE]:
+                                self._child = VideoCaptureElapse(self, self.cfg,
+                                                           self.TopicValues[TP_VIDEOTI],
+                                                           self.TopicValues[TP_VIDSPEED],
+                                                           self._motionEnabled)
+                            else:
+                                self._child = VideoCapture(self, self.cfg,
+                                                           self.TopicValues[TP_VIDEOTI],
+                                                           self._motionEnabled)
                             self._video = True
                             self.publish_state(TP_VIDEO)
                             self.activeThreads.addThread(self._child)
+                            self.TopicValues[TP_REC]=True
+                            self.publish_state(TP_REC)
                         break
                     elif TP_HTTP == tp:
                         self.TopicValues[tp]= payload
-                        if payload == "ON":
+                        if payload == hass.HASS_STATE_ON:
                             self._child = HTTPStreamCapture(self, self.cfg, self._motionEnabled)
                             self._hstream = True
                             self.publish_state(TP_HTTP)
                             self.activeThreads.addThread(self._child)
+                            self.TopicValues[TP_REC]=True
+                            self.publish_state(TP_REC)
                         break
                     elif TP_UDP == tp:
                         self.TopicValues[tp]= payload
-                        if payload == "ON":
+                        if payload == hass.HASS_STATE_ON:
                             self._child = UDPStreamCapture(self, self.cfg, self._motionEnabled)
                             self._ustream = True
                             self.publish_state(TP_UDP)
                             self.activeThreads.addThread(self._child)
+                            self.TopicValues[TP_REC]=True
+                            self.publish_state(TP_REC)
                         break
                     else:
                         logging.warning("unhandled payload" + payload)
@@ -383,22 +475,22 @@ class PiCam2Client (hass.MQTTClient):
                 else:
                     if TP_SNAP == tp and \
                             isinstance(self._child, ImageCapture):
-                        if payload == "OFF":
+                        if payload == hass.HASS_STATE_OFF:
                             self._child.trigger_stop()
                         break
                     elif TP_VIDEO == tp and \
                             isinstance(self._child, VideoCapture):
-                        if payload == "OFF":
+                        if payload == hass.HASS_STATE_OFF:
                             self._child.trigger_stop()
                         break
                     elif TP_HTTP == tp and \
                             isinstance(self._child, HTTPStreamCapture):
-                        if payload == "OFF":
+                        if payload == hass.HASS_STATE_OFF:
                             self._child.trigger_stop()
                         break
                     elif TP_UDP == tp and \
                             isinstance(self._child, UDPStreamCapture):
-                        if payload == "OFF":
+                        if payload == hass.HASS_STATE_OFF:
                             self._child.trigger_stop()
                         break
                     else:
@@ -419,27 +511,38 @@ class PiCam2Client (hass.MQTTClient):
 
         if isinstance(child, ImageCapture):
             self._snapshot = False
-            self.TopicValues[TP_SNAP] = "OFF"
+            self.TopicValues[TP_SNAP] = hass.HASS_STATE_OFF
             self.publish_state(TP_SNAP)
+            self.TopicValues[TP_REC]=False
+            self.publish_state(TP_REC)
             self._child = None
-        elif isinstance(child, VideoCapture):
+        elif isinstance(child, VideoCapture) or \
+             isinstance(child, VideoCaptureElapse) :
             self._video = False
-            self.TopicValues[TP_VIDEO] = "OFF"
+            self.TopicValues[TP_VIDEO] = hass.HASS_STATE_OFF
             self.publish_state(TP_VIDEO)
+            self.TopicValues[TP_REC]=False
+            self.publish_state(TP_REC)
             self._child = None
         elif isinstance(child, HTTPStreamCapture):
             self._hstream = False
-            self.TopicValues[TP_HTTP] = "OFF"
+            self.TopicValues[TP_HTTP] = hass.HASS_STATE_OFF
             self.publish_state(TP_HTTP)
+            self.TopicValues[TP_REC]=False
+            self.publish_state(TP_REC)
             self._child = None
         elif isinstance(child, UDPStreamCapture):
             self._ustream = False
-            self.TopicValues[TP_UDP] = "OFF"
+            self.TopicValues[TP_UDP] = hass.HASS_STATE_OFF
             self.publish_state(TP_UDP)
             self._child = None
+            self.TopicValues[TP_REC]=False
+            self.publish_state(TP_REC)
         elif isinstance(child, PanTiltServoMotors) or \
              isinstance(child, PanTiltStepMotors) :
             self._PanTiltCam=None # delete old reference
+        else:
+            logging.warning("unhandled client down instance")
 
     def motion_detected(self):
         """ callback function when motion has been detected """